app/requirements.txt
fastapi
uvicorn
scikit-learn
pydantic
numpy
google-cloud-logging
prometheus-fastapi-instrumentator